insert conditionals handling into ExpressionSplitter pass
