+++ src/passes/index.ts
@@ -2,11 +2,10 @@ export * from './annotateImplicits';
 export * from './builtinHandler';
 export * from './bytesConverter';
 export * from './cairoUtilImporter';
-export * from './conditionalFunctionaliser';
 export * from './constantHandler';
 export * from './deleteHandler';
 export * from './enumConverter';
-export * from './expressionSplitter';
+export * from './expressionSplitter/expressionSplitter';
 export * from './externalArgumentModifier';
 export * from './externalContractHandler';
 export * from './externalInputChecker';
